Fix MSVC compiler warnings
- silenced narrowing conversion warnings with explicit casts
- fixed a variable shadowing warning by removing an unused variable

+++ src/file_descriptor.cpp
@@ -265,7 +265,7 @@ std::streamsize file_descriptor_impl::read(char* s, std::streamsize n)
 {
 #ifdef BOOST_IOSTREAMS_WINDOWS
     DWORD result;
-    if (!::ReadFile(handle_, s, n, &result, NULL))
+    if (!::ReadFile(handle_, s, static_cast<DWORD>(n), &result, NULL))
     {
         // report EOF if the write-side of a pipe has been closed
         if (GetLastError() == ERROR_BROKEN_PIPE)
@@ -289,7 +289,7 @@ std::streamsize file_descriptor_impl::write(const char* s, std::streamsize n)
 {
 #ifdef BOOST_IOSTREAMS_WINDOWS
     DWORD ignore;
-    if (!::WriteFile(handle_, s, n, &ignore, NULL))
+    if (!::WriteFile(handle_, s, static_cast<DWORD>(n), &ignore, NULL))
         throw_system_failure("failed writing");
     return n;
 #else // #ifdef BOOST_IOSTREAMS_WINDOWS
